修复创建payment返回主键一直是1的问题

+++ cloud-provider-payment8001/src/main/java/com/hruan/springcloud/service/PaymentService.java
@@ -4,6 +4,6 @@ import com.hruan.springcloud.entities.Payment;
 import org.apache.ibatis.annotations.Param;
 
 public interface PaymentService {
-    int create(Payment payment);
+    long create(Payment payment);
     Payment getPaymentById(@Param("id") Long id);
 }
